add getPathName selector
