+++ src/containers/menus/globe-page-indicator/selectors.js
@@ -1,0 +1,7 @@
+import { createStructuredSelector } from 'reselect';
+
+import { getPathname } from 'selectors/location-selectors';
+
+export default createStructuredSelector({
+  pathname: getPathname,
+});
